Added MessageBox for when the mesh fails to load. Removed CreateRight/LeftHanded versions for mesh.  Now there is a single Create function that takes in right-handed indices. fixed minor warnings with included math functions.

+++ Code/Engine/Core/FloatMath.inl
@@ -88,7 +88,7 @@ namespace Engine
 					return i_lhs == i_rhs;
 
 				//Compare the ULP difference to our maxULP difference allowed
-				return abs(left - right) <= maxULPdiff;
+				return static_cast<size_t>(abs(left - right)) <= maxULPdiff;
 			}
 		}
 	}
